Inject Client and System; consolidate mocks
