+++ module/client/lauth/authorizer.h
@@ -1,22 +1,22 @@
 #ifndef _LAUTH_AUTHORIZER_H_
 #define _LAUTH_AUTHORIZER_H_
 
+#include "lauth/client.h"
 #include "lauth/request_info.h"
+#include "lauth/system.h"
 
 #include <memory>
 #include <string>
 
 namespace mlibrary::lauth {
-    class System {
-        public:
-        virtual std::string getHostname();
-        virtual ~System() = default;
-    };
-
     class Authorizer {
         public:
         Authorizer() : system(std::make_unique<System>()) {};
         Authorizer(std::unique_ptr<System>&& system) : system(std::move(system)) {};
+        Authorizer(
+            std::unique_ptr<System>&& system,
+            std::unique_ptr<v1::Client>&& api
+        ) : system(std::move(system)), client(std::move(api)) {};
         Authorizer(const Authorizer&) = delete;
         Authorizer& operator=(const Authorizer&) = delete;
         Authorizer(Authorizer&&) = delete;
@@ -28,6 +28,7 @@ namespace mlibrary::lauth {
         std::string getHostname();
         protected:
         std::unique_ptr<System> system;
+        std::unique_ptr<v1::Client> client;
         bool defaultCleanup = false;
     };
 }
