+++ module/client/test/BUILD
@@ -1,5 +1,5 @@
 cc_test(
   name = "client_test",
-  srcs = glob(["*_test.cpp"]),
+  srcs = glob(["*_test.cpp", "*.h"]),
   deps = ["@googletest//:gtest_main", "//lauth:lauth-client"],
 )
