+++ module/client/test/mocks.h
@@ -1,0 +1,35 @@
+#ifndef _LAUTH_TEST_MOCKS_H_
+#define _LAUTH_TEST_MOCKS_H_
+
+#include <gtest/gtest.h>
+#include <gmock/gmock.h>
+#include "lauth/client.h"
+#include "lauth/http_client.h"
+#include "lauth/system.h"
+
+#include <memory>
+
+using namespace mlibrary::lauth;
+
+class MockSystem : public System {
+    public:
+    MOCK_METHOD(std::string, getHostname, ());
+};
+
+struct StubRawClient : public httplib::Client {
+    StubRawClient() : httplib::Client("http://default.invalid") {}
+};
+
+class MockHttpClient : public HttpClient {
+    public:
+    MockHttpClient() : HttpClient(StubRawClient()) {}
+    MOCK_METHOD(std::string, getBody, (const std::string&));
+};
+
+class MockClient : public v1::Client {
+    public:
+    MockClient() : v1::Client(std::make_unique<MockHttpClient>()) {}
+    MOCK_METHOD(v1::User, getUser, (std::string));
+};
+
+#endif
